Finish all

+++ app/controllers/public/orders_controller.rb
@@ -8,20 +8,23 @@ class Public::OrdersController < ApplicationController
 
   def create
     order = Order.new(order_params)
-    order.save
-    @cart_items = current_customer.cart_items.all
+    if order.save
+      @cart_items = current_customer.cart_items.all
 
-    @cart_items.each do |cart_item|
-      @order_details = OrderDetail.new
-      @order_details.order_id = order.id
-      @order_details.item_id = cart_item.item.id
-      @order_details.price = cart_item.item.with_tax_price
-      @order_details.amount = cart_item.amount
-      @order_details.making_status = 0
-      @order_details.save!
+      @cart_items.each do |cart_item|
+        @order_details = OrderDetail.new
+        @order_details.order_id = order.id
+        @order_details.item_id = cart_item.item.id
+        @order_details.price = cart_item.item.with_tax_price
+        @order_details.amount = cart_item.amount
+        @order_details.making_status = 0
+        @order_details.save!
+      end
+      CartItem.destroy_all
+      redirect_to orders_thanks_path
+    else
+      render :new
     end
-    CartItem.destroy_all
-    redirect_to orders_thanks_path
   end
 
   def confirm
